+++ Calibration/Calibration/CDecodePhase.cpp
@@ -47,7 +47,7 @@ bool CDecodePhase::DeleteSpace()
 // 根据灰度图统计结果
 bool CDecodePhase::CountResult()
 {
-	this->m_result.create(this->m_resRow, this->m_resLine, CV_32FC1);
+	this->m_result.create(this->m_resRow, this->m_resLine, CV_64FC1);
 	Mat SinValue0;
 	Mat SinValue1;
